MakeProfiles builds 3D ellipsoids
3D ellipsoids can now be built with MakeProfiles with any shape and
orientation. This is a major commit unifying all the work that was
necessary to accomplish this feature. In the process, several library
fuctions were also corrected and several bugs were also found and
fixed. Bugs relating to the building of Gnuastro that were reported from
outside of this task were implemented in the master branch until now also,
so there is no need to merge master with this branch. All the major
changes/additions have been described in the NEWS file.