Minor corrections in MakeProfiles manual
I had forgot to describe the `astmkprof-3d.conf' configuration file that is
installed with the rest of the configuration files at install time. It
greatly simplifies the creation of 3D ellipsoids. So a paragraph was added
in the manual describing this configuration file and how to use it.

Also, the `--mode' option was moved to the profiles group of options
similar to the other options that define the interpretation of the catalog
columns.