Noised cube created in make check
During `make check', MakeProfiles creates a 3D datacube with three profiles
(introduced in a previous commit). With this commit noise is added to that
cube with MakeNoise.

I also noticed that the `tests/mkprof/3dcat.txt' was not added in the Git
repository (because of its `.txt' prefix). So it is now version controlled
with this commit.