Some minor corrections in code and comments
The comments above the column counters of `bin/mkprof/ui.c' still referred
to the inverse order of linked lists. This was for a time when we would
directly start reading the list. But now, we use `gal_list_str_reverse'
after setting the column names, so this comment was obsolete and would only
confuse the code readers. Also, the `coordcounter' variable is no longer
used, so it was removed.

Finally, in `README', I noticed that we were missing a `:' after
BuildProgram! This has been added.