Identifiers for integer constants in BZERO comparisons
For each integer width, the FITS standard only defines one type: unsigned
8-bit, signed 16-bit and so on. So to allow other types (for example signed
8-bit or unsigned 16-bit), it defines constant values for BZERO. The
constants for 8-bit and 16-bit integers have no particular problem, but for
the 32-bit and 64-bit integers, if there is no identifiers, the compilers
may produce warnings (on 32-bit and 64-bit systems).

Until now, only the 64-bit value was given an identifier of `.0f' to read
it as float. But floats (even double precision), don't have the necessary
resolution for such large integers (`double' can only accurately keep an
integer until 15 significant decimal digits). For the 32-bit value, I got a
compiler warning on a 32-bit `armv81' GCC ("Warning, this decimal constatn
is unsigned only in ISO C90").

To fix the problem, the `LU' and `LLU' identifiers were added for the
32-bit and 64-bit constants of `fits_type_correct'. However, this doesn't
fix another problem with comparing a floating point to the 64-bit integer
constant. In a check, I noticed that after defining `a' and `b' as: `double
a=9223372036854775808LLU, b=9223372036854775807LLU;' (note that `a=b+1'),
the condition `a==b' succeeds (because of limited floating point
precision). So I just sent a mail to William Pence (author of CFITSIO)
about the issue to get his comments.