Bug fixes in master: commits 1ff1c25 and 540af65
Two recent bugs were found in Gnuastro and fixed in the master branch. With
this commit, they are also copied/implemented here. I didn't actually use
Git's merge because the history can get very messy with all the merges with
master. It is cleaner to manually copy the files like this.