NoiseChisel's convolution step in 3D completed
Work on adapting NoiseChisel to 3D is ongoing and with this commit, a major
block has been removed: 3D spatial convolution is now tested and and
working on 3D data. So NoiseChisel is now ready to start doing its
thresholding on the images.

To do the job, a default 3D kernel has been defined with this commit as a
separate header that is included into the `ui_prepare_kernel' function (it
was too large to be put manually into the full source and would make
parsing by eye hard).

In the process, the following issues have also been synchronized with the
master branch:

 - A bug found by Vladimir in the Warp program is now implemented in this
   branch also.

 - The corrections and bug fixes in the tessellation and convolution
   libraries of this commit were also implemented in the master branch.