New --mcolisbrightness option for MakeProfiles
It might happen that the user knows the total brightness of the profiles
they want and don't want to go through the trouble of converting the
magnitude to brightness (sum of pixel values). With this commit, they can
do this by using the new `--mcolisbrightness' option.

Also, I noticed that the configuration file to help use NoiseChisel in 3D
was not installed in the previous commits. With this commit,
`astnoisechisel-3d.conf' is also installed.

Finally, a notice was added in the book on how to define a shell alias for
MakeProfiles and NoiseChisel to automatically call them along with their 3D
configuration files.