Default NoiseChisel kernel in 3D changed to 1.5 pixels
Until now the default NoiseChisel kernel in 3D had a 2 pixel FWHM in the
first two dimensions (similar to the 2D case). However, this mixes the 3D
dataset's voxels (smooths) too much. Note how here, the shrinking of the
dynamic range (and possibly skewing the distribution) of the dataset
(signal+noise) is the primary factor. I currently feel that sampling is
only important as it relates to the ratios between the different
dimensions.

I also noticed that in the manual, there was no discussion on the default
3D kernel properties, so they were added.