More --coordcol options acceptable in Crop
Until now, in Crop, `--coordcol' had to be called exactly the same number
times as the input's dimensions. This was because it was being used
directly in the function to read the columns. This is inconvenient (for
example it didn't allow using `--coordcol' in default configuration files).

With this commit, if it is called more than the input's number of
dimensions, there is no problem. The extra columns will be deleted before
being used. Crop aborts with an error only if there is not enough
`--coordcol' calls.