No 3D projections in function to inspect NoiseChisel outputs
Until now, the two 3D frames that opened with the rest of the 2D slices
were projected by 45 degrees to clearly show their 3D nature. But after
using it a while, I noticed that atleast on IFU data that I am testing
with, it is easier to not change the viewing angle and only look at the 3D
dataset with azimuth and elevation of zero. This is because it will be the
same size as the rest of the frames and is a superposition of all the 3rd
dimension information. When the fields are even slightly crowded, the 45
degree viewing angle is not useful at all and is even confusing. If the
user wants to, they may rotate the cube in any way they like.

Also, a recent change in master (commit a8422ee3: "Write access only when
necessary in Fits program") is implemented in this branch with this commit
(just one line).