Oversegmentation connectivity one less for 3D
In 2D, when looking at the neighbors, we use the dimensionality for
connectivity (or 8-connected neighbors). In 2D this makes 4-connected
rivers (more pixels as rivers). This is fine in 2D, but in 3D, it can mean
a circumference in each slice. So through this change, we are making the 3D
rivers slightly thinner. This apparently gives better 3D clump results.

In the process of inspecting the clumps, I noticed that in
`clumps_get_raw_info' we are not finding/using the 3rd dimension
information and un-initialized values were being used instead. Now it finds
the clump center in 3D and uses the three values to find the tile to return
the standard deviation.

I also noticed that in `clumps_true_find_sn_thresh', we have to use
`GAL_TYPE_INT32' directly instead of relying on `clprm.snind->type'. This
is because of a recent correction in `clumps_make_sn_table' where we stop
continuing when `cltprm->numinitclumps==0' (so `clprm.snind->type' was not
initialized).