Single to conditionally open ds9 for multi-HDU 2D or 3D
Until now, there was no elegant way (a single command) to open ds9 for
multi-extension 2D or 3D inputs both on the GUI and in the command-line. So
using Gnuastro's FITS program, a script has been added for the job. The old
shell functions for 3D viewing have been removed from the NoiseChisel
output section and it is fully now described in the "Viewing multiextension
FITS images" appendix.