MakeProfiles kernel option can be elongated in 3rd-dimension
Until now, when using the `--kernel' option, it was only possible to have
the third dimension be at most as extended as the first two. However,
during my tests on MUSE datasets, cases came up that it would be useful to
have the kernel extended in the third dimension more than the first two.

With this commit, the last value of the 3D kernel profiles can be a value
larger than 1 and in such cases, the kernel will be elongated in the third
dimension (bullet-shaped). The description of this option for the book has
also been modified.