Connectivity of 2, for filling pseudo-detection holes in 3D
Until now, the number of dimensions (strongest connectivity) was used to
fill holes in the pseudo-detections. For 2D, the strongest connectivity (2)
is fine. But for 3D, the strongest connectivity is too strong and will not
be too useful for diffuse signal. So for 3D, we are now using the
second-strongest connectivity of 2 (numerically: same as the 2D case).