Correct checks for kernel dimension in NoiseChisel and Segment
Until now, NoiseChisel wasn't checking if the dimensionality of the kernel
is the same as the input. Also, Segment was checking the dimensionality
after possibly setting `p->kernel=NULL' (when `--kernel=none' is called,
which caused a segmentation fault).

With this commit, we now check the dimensionality of the kernel in both
NoiseChisel and Segment. In both cases, this check is also done only when
`--kernel' has not been `none'.