Narrow-band areas from MakeCatalog in 3D datasets
When working with 3D datasets (especially IFU datasets), it is commonly
necessary to get narrow-band information for the objects/clumps. With this
commit the steps have been done for doing this.

Currently it is only for the area, but later, we will be able to do
elliptical measurements and etc also.