New spectrum measurement from 3D inputs in MakeCatalog
MakeCatalog can now estimate the spectra of 3D detections using the
`--spectrum' option. This was implemented by defining an extra step after
parsing of objects (and finding the label's 2D projection map). The output
spectra are written into separate tables (one for each label).

The spectra for each label are writte into a list of datasets which are
held together using an array of datasets during the processing. Once
processing is finished, they are written into the output tables. Maybe
later we can add an option to write the tables using a separate thread
using CFITSIO's low-level functionality) to speed it up (similar to how we
write the separate profiles of MakeProfiles into the final image: with each
thread giving their output to a separate thread that is devoted to the
writing). But for now, I don't have time to implement that.