Projected spectra given NaN when no measurement on slice
In the spectra extensions of MakeCatalog's output we report the area of
measurements on each slice for all spectra except the projected
spectra. But this caused complications in interpretting the value of zero
in a projected spectra: is the sum of values zero, or is the whole slice
blank (for example because of Sky lines)?

So with this commit, when no measurement could be made in the projected
spectra, at the end of parsing the slice, we set it (and its error) to NaN.