All MakeCatalog spectrums set to NaN when no measurement was done
Because zero is meaningful, recently, in the projected measurements, we set
the slices were there was no measurement to NaN. But the same argument also
applies to the `SUM', `SUM_ERR', `SUM_OTHER', and `SUM_OTHER_ERROR'
columns. So in this commit, when there are no labeled voxels in a slice, it
is set to NaN in all its spectra.