NoiseChisel: independent test of input's dimensions
Until now there was no independent test of dimensionality after NoiseChisel
read the input dataset. As a result, when the input's dimensions were not
acceptable, it would crash with a suggested bug in the `ui_ngb_check'
function.

With this commit, immediately after reading the input, NoiseChisel checks
is dimension independently. So the bug report of `ui_ngb_check' is
reasonable.