astscript-radial-profile: adding script for making radial profiles
With this commit, I am adding a script for making radial profiles. This
is the very early version of this script, and because of that, a lot of
information (specially for the book) is missing and has to be added.
However, I am commiting this work to be safe and not loose information.

Basically, this script takes an input image and compute the radial
profiles over defined apertures. To do that, the apertures are defined
with `astmkprof' setting the center of the apertures (WCS o IMG
mode) as well as the ellipticity and the position angle. In the same
way, the minimum and maximum radius have to be selected (otherwise it
will take the maximum possible value for the radius). Once the apertures
are defined, it uses `astmkcatalog' for obtaining the wanted measure
(tipically mean, median, brightness, etc.) over the defined apertures.

So, at the end, the result is a table with two columns, first one is the
radial distance from the center, and the other column is the measured
parameter over the aperture.