astscript-radial-profile: added the option -b for binning the data
Until this commit, the radial profiles were obtained in a pixel scale.
That is good for high resolution but sometimes it is necessary to have a
bigger bin size in order to increase the signal to noise ratio
(specially as far as you go from the center).

With this commit, this has been solved by averaging the already obtained
radial profiles in bins of a given size. To do that, a small `awk' code
has been included. Now, the script reads the bin size with the parameter
`-b', and gives as output a binned radial profile. Based on this post:
https://stackoverflow.com/a/30353074