astscript-radial-profile: allowing to manipulate the final profile
Until this commit, the only possible manipulation for obtaining the
final radial profile was the binning of the data. However, it is
convenient to have the possibility of modify the radial profile as the
user want. For example, scaling the radius column to obtain them in
arcsec instead of pixels. The same for the second column: it could be
necessary to obtain surface brightness magnitudes instead of the raw
measure done until now.

With this commit, it has been added the possibility to manipulate the
two columns of the radial profile. There is just necessary to specify
the wanted operation with the options --X and --Y (--xarith and
--yarith) from the command line. The syntaxis and notation for
manipulate the columns is the same used by Table when arithmeting
different columns. If nothing is wanted to be done, then the data (both
columns) is just multiplied by 1.