astscript-radial-profile: fixing some bugs and modifying metadata
Until this commit, the radial profile script was some bugs that I found
once I started to check it from scratch. For example, the axis ratio
(`-Q' or `--qratio') was set to 0 as default value, however, it cannot
be zero! The default value should be 1. Another important bug fixed is
the generation of an output radial profile even if the user do not
specify the name with the `-o' or `--output' options.

With this commit, these bugs have been fixes. In addition to this, the
temporal files generated for obtaining the final radial profile have
been also modified to have better names. The metadata information
included in the final radial now contains the operations that have been
done in order to obtain both columns. So, it is kept the binning and the
final arithmetic operation as comments.