astscript-radial-profile: adding documentation to the GNUASTRO book
Until this commit, any information about the `astscript-radial-profile'
script was written into the gnuastro book. This is simply because it was
being developed. Now that I have been testing it with a couple of
examples and it is more o less mature, it is necessary to add the
documentation and information to the book.

With this commit, only the introduction and a couple of examples about
the script have been added to the gnuastro manual book. The description
of all different options is still remaining.