astscript-radial-profile: user can now specify output column names
Until this commit, the output column names were `radius', and `op',
where `op' was the operator used for computing the measured values over
the different apertures (`--measure=op'). However, when trying to
operate with the output table (in which the column names were `radius'
and `median'), Table program crashed. My guess is that it was because
the column name `median' was interpreted as the operator `median'. But,
in any case it is good to be able to specify the output column names of
the radial profile table.

With this commit, two parameters (`--iname' or `-i', and `--jname' or
`-j') have been added. They correspond with the output column names for
the radius and measured values, respectively. If the user does not
specify the name of the first column (`-i'), the name will be `radius'.
If the user does not specify the name of the second column (`-j'), the
name will be the operator (`--measure') with the letter `v' at the
beginning.  For example, if the operator for computing the measured
values is `--measure=mean', the column name will be `vmean'.