astscript-radial-profile: setting name and version from the compilation
Until this commit, the version number and the name of the script were
set by hand (inside of the script). However, this does not allow to have
an automatic upload of the version number when changing the version
number of Gnuastro (because of other commits).

With this commit, both parameters (version number and name of the
script) are now filled in the compilation of the script.