astscript-radial-profile: removing --rmin option because it is not used
Until this commit, there was the option `--rmin' for setting the minimum
radius for the radial profile. This option comes from the very first
version of the script, when it used a for-loop over the different radius
for making the profiles. However, in the current version it is using the
option `distance' for generating profiles with the radial distance from
the center (with `astmkprof'). As a consequence, this option is not used
anymore. With this commit, this option has been removed.