astscript-radial-profile: modification to use $ instead of c in Table
Until this commit, when refering to columns in `asttable', it had to be
specified by typing `c1', `c2', etc. However, with the recent changes in
Gnuastro, now it is `$1', `$2', etc. This has been introduced with the
aim of be more general and follow the standard notation like `awk'. With
this commit, this new convention has been introduced in the
astscript-radial-profile. In the last step it invokes `asttable' to
operate the columns, so it was necessary to change the call to the
columns. Also important is to note that the operator now are enclosed
between double quotes to avoid problem with the white spaces.

Thanks to Mohammad and Samaeh for helping in solving this bug!