astscript-radial-profile: sigma clip parameters for measurements added
Until this commit, only default values for sigma clipping measurements
were considered (default parameters assumed in Catalog). However, if the
user wanted to measure with another sigma clipping paramter it was not
possible! With this commit, this problem has been solved. Now, the
radial profile script takes the parameter -s or --sigmaclip to be two
floats (number of sigmas and tolerance). The values to be measured will
be computed with these values for sigma clipping operators.