Radial profile script modifications
With this commit, some modifications to the radial profiles script have
been done. A temporary directory was defined for doing the process on it
and prevent occupying the source directory.  Also, the cropped image will
be created based on central parameters and maximum radius to continue the
process on it instead of the whole input image.

Thank Mohammad for useful advices and his modifications.