NoiseChisel: Sky estimation on tiles used for qthresh
Until now, the Sky was independently calculated after the detection
mask. However, we already have a robust tile outlier removal in the
'qthresh' step. Infact the qthresh tessellation just before interpolation
is much more resiliant to residual diffuse flux compared to just estimating
the sky from the fraction of undetected pixels in a tile.

Therefore with this commit, we now save the good tiles (as a blank flag
image) just before interpolating over the qthresh tessellation. Later, when
estimating the sky (both for the pseudo-detections and the final
detections), we only check the Sky value (and the fraction of undetected
pixels) only on the good tiles found above.

This results in a major improvement in NoiseChisel's final Sky image, with
a much smaller footprint of possibly undetected signal.

In the process, I also went through the two tutorials and adjusted anything
that needed it. I also added a new section showing how we can make 2D
histograms.