Corrected segmentation fault, included aperture
The cause of the segmentation fault was in 'match_kdtree_worker': the
'point' pointer was never given a value! In effect, the returned value to
the call to 'gal_pointer_allocate' (that was meant to be written in
'point') wasn't written anywhere.

As a result, the 'printf' statement (and later in the loop where we filled
the point's value) were trying to write into a 'NULL' pointer (the
initialized value of 'point'). So fixing the segmentation fault was easy:
We just had to put a 'point=' before the 'gal_pointer_allocate' ;-).

After fixing this, I thought it may be good to add the aperture checking
step here too: after finding the nearest neighbor through the k-d tree, we
should check if that nearest neighbor is acceptable through the values
given to the '--aperture' argument on the command-line. See the
documentation of this option for more (it may be a circle or ellipse).

To do the aperture checking, I used the same static fuction as the classic
matching ('match_coordinates_distance'). In order to use this, I also
needed to use the static 'match_coordinates_sif_prepare' function which
does all the preparations for apperture matching. If the elliptical
distance of the nearest neighbor is not within the acceptable aperture,
then the written index for that row will be 'GAL_BLANK_SIZE_T'.

So with the '--aperture=1' value of 'tests/during-dev.sh', only one of the
points will currently have a non-NaN match. To get more matches, increase
the value to the '--aperture' option. You can also try the more complex
aperture values and see the elliptical matching in effect.

In the end, after the parallelized matching, the 'gal_match_kdtree'
function will simply print the matching coordinates. You can use this as a
model to continue with the rest of the steps.

Try to make the output similar to the same format as the classic output, so
you can easily leave the rest to be written by the existing functions and
not worry about creating the log file, or sorting rows/columns and etc.