Library (pointer.h): unused RAM decreased to 250MB
Until now, the minimum amount of RAM in the system that Gnuastro programs
and libraries would leave free was 750MB. However, on a recent build in
Debian's tracking system, running of MakeNoise failed in 'make check' on
the Hurd kernel (which is still based on a 32-bit address space and thus
has a very limited RAM of 830MB). Initially I thought this crash was caused
due to parallel operations of that Debian Hurd system.

With this commit, the minimum available memory that will never be used by
Gnuastro was set to 250MB, which is still more than enough for basic system
operation (given that the system was already occupying its necessary RAM
before the Gnuastro program was running). This is only for operations in
parallel to the program.

But after the fix, I noticed a more simplier cause of this problem: instead
of passing 'EXIT_SUCCESS' to the 'error' function (so it doesn't crash the
program, but just prints a warning), I had passed 'EXIT_FAILURE' after the
warning of not finding the 'MemAvailable' keyword.

Also, in the process, I done some leak-checking in MakeNoise to make sure
there was no memory issues there and found and fixed a few leaking
cases. Afterwards, I also done a similar leak test on NoiseChisel (without
any parameters) and fixed a few leaks there too.