Query: default dataset, using random string when no --output
Until now, the query program required the '--dataset' arguments. However,
in the case of Gaia (and possibly many other databases), it is safe to
assume a certain dataset as default (to use when people don't explicitly
ask for a dataset).

Another issue with the Query program was that when no '--output' was given
by the user, Query would calculate the DATASUM of the downloaded table and
use that as the default name. But this had multiple problems: 1) only
worked for FITS tables, 2) needed extra processing, 3) Needed full download
of the dataset before assigning a name, 4) would right over a previously
downloaded file with the same DATASUM (differing metadata), and etc.

With this commit, a default dataset is assumed for Gaia ('edr3'), allowing
users to run Query, without having to give '--dataset=edr3' if this is what
they want (which is the most probable scenario). In this case, Query will
first print an extra line saying that which dataset it will be using.

To address the DATASUM issue, when no output name is given, Query simply
use a random 6-character string and append it to the name of the database
to create the default FITS file.

Also, Query, now prints the dimensions of the downloaded table just before
printing the final statement and finishing.

The issue with DATASUM was recommended by Francois Ochsenbein.