Central coordinates of the cropped image
Until now, to do an efficient process, at the beginning of the script, we
cropped the input image based on the specified coordinates and maximum
radius and then continue the process on the cropped image.

With this commit, the definition of the central coordinates after cropping
the image changed. Whit this change, the script be usable for the star in
the image's border, and in a case which we have the half star to calculate
radial profile.