Query: new --keeprawdownload option
Until now, the Query program would read the raw downloaded table, read it
into memory and write that into the output file with CFITSIO. Giving all
output tables the same format, irrespective of which database they are
downloaded from and also un-compressing the data in case the server gives a
compressed FITS file, like bug #59469).

In the end (after internally processing the raw downloaded file), query
would delete the raw downloaded file. However, the raw downloaded file can
possibly have important metadata that the user may need (for example the
version of the database version and etc).

With this commit, the new '--keeprawdownload' option has been added to
Query for this purpose: when given, the raw downloaded file will not be
deleted and it will remain in the same directory as the output.