Query: Enabled access to VizieR
Until now, Query only connected with ESA's Gaia database. But this was only
for a proof of concept. Therefore with the help of Francois Ochsenbein, I
learnt the similar 'curl'-based TAP access to VizieR and was able to
implement it in this commit.

In short, the old 'gaia_query' function is now renamed to 'tap_download'
which is placed in a new file called 'tap.c'. The initial sanity-checks and
preparations for Gaia have been kept in 'gaia.c' and a new 'vizier.c' file
has been added with similar checks for VizieR.

Also, a feature has been added that databases can be accessed through
various URLs (mirrors): by storing the URLs as a simply-linked list of
strings. If the 'curl' command is not successful in one URL it will go to
the next. Gaia only has one URL, but from the 'vizquery' source, I saw that
VizieR has many mirrors in many countries so they are added for
Vizier. However, when I tried them, the 'curl' command failed to connect to
them. So they are commented now. I am in touch with Fancois to see how this
should be fixed.