Statistics: correct dash in CTYPE in 2D histogram image's WCS
Until now, when building the 2D histogram as a FITS image, the raw name of
the column would be written into the 'CTYPE' keyword (specifying the name
and type of the coordinate). However, in the WCS standard, a dash (-) is
meaningful (it distinguishes between the name and projection type, for
example 'RA---TAN' in most astronomical images).

Therefore, when the column name contains a dash that doesn't fulfill the
WCS standard's convention for CTYPE, WCSLIB will correctly complain and
abort.

With this commit, the column name is now checked and if it contains a dash,
the dash will be converted to an '_' before righting into CTYPE. Ofcourse,
the original column name will not be touched.

During the tests, I also noticed two issues:

 - We were not deleting the output files if they already existed, possibly
   causing confusion. So a check has also been added for this.

 - When the number of bins in each axis is different, the values weren't
   being written properly, and resuling in a wrongly placed 2D histogram
   image. This was due to ignoring the difference between the C and FITS
   orders.

 - To have a square 2D histogram by default, the default value of
   '--numbins2' was set to be the same as '--numbins'.