Javascript license information updated for LibreJS
I noticed that the Gnuastro manual has JavaScript weblabels. This is one of
the rare examples I know of, so looked at it with LibreJS enabled (version
7.20.2). It worked fine for some html nodes, but the single-page version
gave an error.

LibreJS was complaining that jax.js and BasicLatin.js didn't have known
licenses. So I added weblabels for those to javascript.html. Then LibreJS
kept asking for more. To quiet it down, I needed to add 6 new labels.