Query: log file (in case of error) is saved in output directory
Until now, when the download didn't succeed for any reason, the log-file
(containing the cause of the error from the server) would be written in the
running directory, even when the user requested the output to be written in
another directory.

But this is not good because the 'log' file is also an output and when the
user asks to build the output in a certain directory, the log file should
also be built there. This also applied to the temporarily downloaded file!
But this is very bad, because if it is a large file, this can be
problematic.

With this commit, this issue has been fixed and the temporarily downloaded
file (and also the log file if there was an error) is stored in the output
file.

In the process a few other issues have been fixed:

 - In Query, when the '--quiet' option is given, the '-s' option (for
   silent) is also given to 'curl' so nothing is printed.

 - The option-parsing functions didn't have a check in case the value was
   an empty string (for example when the value is a shell variable that is
   empty). So such a check has been added.