Book: minor edit in section on dependencies in package managers
Antonio Diaz Diaz reported a typo in the part that the packaging of
Gnuastro in package managers was discussed (an 'is' was mistakenly written
as and 'if').

With this commit, that has been fixed. But after reading though the section
once, I thought it is better to include a link to Repology to actually show
this to the readers: show them how different package managers have
different versions! Because of this, I added this discussion after the
first point (on the differing, and low, versions in some systems).