Query: can now get dataset information, and other new features
Until now it was necessary to know the dataset information (within the
database) a-priori (before running Query). In the case of Gaia, this wasn't
a major problem because it isn't too large with one dominant
dataset. However, in the case of VizieR (with tens of thousands of
datasets), this was very annoying! Fortunately Francois Ochsenbein helped
alot in this aspect and gave me the necessary commands for obtaining this
information and I was able to generalize them.

With this commit, the necessary low-level changes in Query have been made
to allow printing the information of all datasets within a database, or
columns within a dataset. Therefore, it is now easy to first query the
database and get the list of datasets, then query the dataset and get the
list of columns, and then extract the exact data you want (not necessarily
all the columns). Also, it is now possible to specify which columns to use
for coordinate matching. The newly added options to Query are:
  --ccol: to specify the coordinates used for RA and Dec.
  --information: to get dataset and column information.
  --limitinfo: to constrain the printed information.
  --head: only print top rows.

In the process, the documentaion of Query was also fully edited and mostly
re-written to reflect the new changes. The following changes were made in
other parts.

 - In the FITS library, two new functions have been added to allow the
   programs to add COMMENT keywords to the output:
      - gal_fits_key_list_comment_add
      - gal_fits_key_list_comment_add_end
   These were necessary due to the new feature to write the full
   downloading command into the header of the downloaded file. Also the
   function to write keywrods ('gal_fits_key_write_in_ptr') accounts for
   writing comments.

 - In the internal option library, the new 'gal_options_merge_list_of_csv'
   function has been added to merge a list of strings (where each may be a
   CSV string) into one list with all values separated by commas being
   sparate.

 - A bug in 'gal_table_print_info' was fixed: when the input has no
   columns, it won't crash with a segmentation fault, it will just return
   (because there is nothing to print).