Library (wcs.h): correcting for unrecognized angstroms units
Until now, immediately after reading the WCS structure from the file
header, we would pass it to WCSLIB's 'wcsfix' function. However, this
function doesn't recognize the 'angstroms' units, only 'angstrom' (note the
differing 's')! As a result, it would hit a segmentation fault within
WCSLIB and the program would crash.

With this commit, we now check the value to the 'CUNIT3' keyword. If its
value is 'angstroms' (not case sensitive), we manually remove the 's' and
this fixes the issue in WCSLIB.

This issue was found with the help of Alireza Molaeinezhad.