Query: several highly used VizieR datasets given short-names
Until now in VizieR we just had a short name for Gaia! However, it contains
many more useful datasets that are also commonly used.

With the help of Francois Ochsenbein, I was able to get the list of the
most-used datasets in VizieR and created a list of about 15 datasets that
were in the top of the list (mostly ignoring older versions when newer
versions existed).

Here is the description of Francois, on how to get the most commonly used
VizieR datasets:"If you look at METAcat
(http://vizier.unistra.fr/viz-bin/VizieR?-source=METAcat) there is a column
named "popu" which contains a number which is incremented when the catalog
is used; therefore the most popular catalog can be retrieved by entering a
condition on "popu" and asking to sort the result by this number ("Sort"
column checked), and check the "Output Order" - checkbox. Then clicking the
"Submit" button will give the result".