Table: cleaned memory managemend in arithmetic operators
Until now, the names and column metadata of each column was re-written for
all the rows! Because the loop for parsing the rows was outside of the
'case' system.

With this commit, this issue has been fixed by bringing the loop within the
'case' structure. While doing the memory checking, I also noticed that we
weren't cleaning the list of arithmetic operators! That has also been fixed
here.