Query: new --noblank option
Until now, if we didn't want rows that have a blank value in a certain
column, we needed to use the '--range' option which was annoying (because
you had to set very large values and generally know the rough value range
of the column.

With this commit, Query now has a '--noblank' operator that like '--range'
is modeled on the similarly named operator in Table. When '--noblank' is
given, it will use the 'XXXXX IS NOT NULL' structure in ADQL to achieve its
goal and let the user simply ask for non-blank rows in a certain column
without worrying about the value range.