Query: value ranges are defined before positional
Until now while constructing the ADQL query, I first had the spatial range
component (selecting rows within a certain distance of RA and Dec for
example), then the value range component (like '--range' or
'--noblank'). But it seems more logical to first limit the rows to those
that are with a certain value (which is computationally much more easier),
then do the spatial checks.

With this commit, this has been done: we now first limit the rows by their
indiviual values, then do the 2D searching.

Also, I noticed that on VizieR, the Gaia DR2 catalog doesn't have RAJ2000
and DEJ2000. So the default spatial columns are changed to 'ra_epoch2000'
and 'dec_epoch2000'.