Query: added support for ASTRON's VO database
With this commit, Query users can now acess ASTRON's VO service with the
'astron' database name. So far, I have just chosen a set of tables for
short-name based on a general reading of the description of the
datasets. However, I look forward to correcting them or adding more.

This task and necessary links were provided by Javier Moldon.