Query: new option to sort downloaded data by certain column
Until now, the order of the retrieved datasets was determined by the server
that provided the data. But this was annoying when a certain order is
needed (an extra call to 'asttable' would be necessary). However, ADQL does
have features to support sorting on the server.

With this commit, the ADQL feature of 'ORDER BY' has been implemented as
the '--sort' option on the command-line.