Book: minor editing to help in readability
Of the sentences in the book started with the number '20' and the previous
sentence ended in a '4'! This structure was hard to read and could cause
confusions.

With this commit, the '20' is changed to 'Twenty'. This helps visually
separating the sentences.

This issue was reported by Thérèse Godefroy.