Query: removed short names for lofartier1 and mvf in ASTRON
Until now, these two datasets were given a short-name, however "lofartier1"
is a catalog of images not cross-matched astrophysical sources, and as
Javier Moldon suggested, "mvf is not very relevant, as it is a verification
field".

With this commit, they have been removed from the short-name list, but are
still accessible by directly using the long name. I also done some small
edits in the book's "Available datasets" section.