Query: corrected check on full Gaia catalog download
Recently a new check was added for Gaia so if no limits are set on the
requested rows, an error is printed that tells the user why the download is
going to fail. However, in this check, I had forgot to include
'--overlapwith' as one of the ways that rows can be limited.

With this commit, the '--overlapwith' option has also been added to one of
the checks so when it is called by the user, this error won't abort the
program.

This issue was reported by Carlos Morales Socorro.