Final output for kdtree matching
The final output is now being produces for kdtree
matching.`gal_kdtree_nearest_neighbour` alredy finds the smallest
distance between the 2 points so there was no further need to calculate
it. I kept track of all the distance by using a new variable in
`match_kdtree_params`.

To remove the multiple matchings, I used the arry `multiple _match` n
`match_kdtree_params` and used it further to check if the same point has
been checked before in `gal_match_kdtree_output`.

The `gal_match_kdtree_output` works similar to that of normal coordinate
matching, but instead of using a permutaion array, simply uses the
matched points to assign values/indexes to the output.

The error in valgrind is still present but produces the correct fits output.
When using valgrind with option '--track-origin=yes', they track back to
malloc in pointer.c. The output is shown using valgrind but due to the
conditional jumps, are not produced normally.