Query: new --dry-run option to just print command, not run it
Until now, the only way a user could check the download command was to
actually run it and then cancel the run to see what is actually being
run. This is very annoying!

With this commit, Query now has a '--dry-run' option precisely for this
purpose. With this option, it will just print the command that it would run
otherwise and abort.