Library (pointer.h): Memory-mapped files no longer hidden
Until now, the memory-mapped files were created in a hidden directory or as
hidden files. However, since this directory can host very large files, it
can be very confusing for the user when the program crashes and the files
aren't deleted automatically.

With this commit, the directory hosting memory-mapped files is
'gnuastro_mmap' (not '.gnuastro_mmap'). Also, when the directory can't be
made and the memory-mapped files have to be written in the running
directory directly, they will be written as 'gnuastro_mmap_XXXXX' (instead
of '.gnuastro_mmap_XXXXX').

In parallel to this, I was also doing spell-checking on the newly added
parts to 'gnuastro.texi', but forgot to commit them separately. So this
commit also includes a spell-check of the book in the new parts.